backend-python/requirements.txt
# requirements.txt

rank_bm25  
langchain-cohere   
langchain-huggingface  

sentence-transformers  
torch  
langchain  
langchain-community  
faiss-cpu  
python-dotenv 
langchain-google-genai
requests  
beautifulsoup4
Flask
flask-cors
gunicorn 